# api/swagger.yaml
---
swagger: "2.0"
info:
  version: "1.0.0"
  title: "hivent-resources-api"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
tags:
- name: "resources"
paths:
  /resources/{id}:
    get:
      tags:
      - "resources"
      operationId: "readResource"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Lecture réussie."
          schema:
            $ref: "#/definitions/Resource"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Resources"
    put:
      tags:
      - "resources"
      operationId: "updateResource"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "resource"
        required: true
        schema:
          $ref: "#/definitions/Resource"
      responses:
        200:
          description: "Mise à jour réussie."
          schema:
            $ref: "#/definitions/Resource"
        400:
          description: "Erreur lors de la mise à jour."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Resources"
    delete:
      tags:
      - "resources"
      operationId: "deleteResource"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        204:
          description: "Suppression réussie."
        400:
          description: "Erreur lors de la suppression."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Resources"
  /resources:
    get:
      tags:
      - "resources"
      operationId: "findResources"
      parameters:
      - name: "type"
        in: "query"
        description: "Type [ ROOM, EQUIPMENT ]<br>TypeWorkGroupEnum"
        required: false
        type: "string"
        enum:
        - "ROOM"
        - "EQUIPMENT"
      - name: "available"
        in: "query"
        required: false
        type: "boolean"
      - name: "beginDate"
        in: "query"
        required: false
        type: "string"
        format: "date-time"
      - name: "endDate"
        in: "query"
        required: false
        type: "string"
        format: "date-time"
      responses:
        200:
          description: "Lecture réussie."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Resource"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Resources"
    post:
      tags:
      - "resources"
      operationId: "createResource"
      parameters:
      - in: "body"
        name: "resource"
        required: true
        schema:
          $ref: "#/definitions/Resource"
      responses:
        201:
          description: "Création réussie."
          schema:
            $ref: "#/definitions/Resource"
          headers:
            Location:
              type: "string"
              description: "URI de la ressource créée"
        400:
          description: "Erreur lors de la création."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Resources"
  /resources/{id}/available:
    get:
      tags:
      - "resources"
      operationId: "isResourceAvailable"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - name: "beginDate"
        in: "query"
        required: false
        type: "string"
        format: "date-time"
      - name: "endDate"
        in: "query"
        required: false
        type: "string"
        format: "date-time"
      responses:
        200:
          description: "Lecture réussie."
          schema:
            type: "boolean"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Resources"
  /resources/{id}/invitations:
    post:
      tags:
      - "resources"
      operationId: "createResourceInvitationAsync"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "calendarEventId"
        required: true
        schema:
          type: "string"
      responses:
        201:
          description: "Création réussie."
          headers:
            Location:
              type: "string"
              description: "URI de la ressource créée"
        400:
          description: "Erreur lors de la création."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Resources"
  /resources/{id}/invitations/{calendarEventId}:
    delete:
      tags:
      - "resources"
      operationId: "deleteResourceInvitationAsync"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - name: "calendarEventId"
        in: "path"
        required: true
        type: "string"
      responses:
        204:
          description: "Suppression réussie."
        400:
          description: "Erreur lors de la suppression."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Resources"
definitions:
  Resource:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Identifiant.Lecture seule.\n"
      name:
        type: "string"
      mail:
        type: "string"
      type:
        $ref: "#/definitions/Type"
    example:
      mail: "mail"
      name: "name"
      id: 0
      type: {}
  Type:
    description: "Type"
    enum:
    - "ROOM"
    - "EQUIPMENT"
  ResourceStatus:
    description: "ResourceStatus"
    enum:
    - "FREE"
    - "NOUVELLE_VALEUR"
  ReponseErreur:
    type: "object"
    properties:
      erreurs:
        type: "array"
        items:
          $ref: "#/definitions/ReponseErreur_erreurs"
  ReponseErreur_erreurs:
    properties:
      champErreur:
        type: "string"
      codeErreur:
        type: "string"
      libelleErreur:
        type: "string"